Adversarial review fixes (17 findings), hostcall builtins, Base corpus
Review workflow (2 reviewers -> 17 verified findings -> fixes + regression
tests, all suites green):
- WasmtimeRunner: refcounted externref root table with finalizer-based
  release (leak verified fixed via RSS), read(WasmMemory) returns owned
  copy (was use-after-free), per-Store ReentrantLock with documented
  concurrency contract, funcref round-trip, robust host trampoline,
  loud anyref/exnref handling, prefer vendored v45 over registered v39 JLL.
- WasmCodegen criticals: sub-word checked sdiv typemin/-1 now traps;
  Char uses raw-bits representation end to end. Majors: shift-count
  saturation before i32 wrap, signed-interpretation ops on unsigned
  sub-word reps (emit_signext!), const-field setfield! traps, Bool
  results normalized to {0,1}, typed select for GC refs. Documented
  latitudes: muladd fusion, unsafe_trunc saturation.

Plus: hostcall mechanism (builtins on host-resident values route through
auto-generated imports: sizeof(::String), === on externref),
getfield(::MemoryRef,:mem), offload thunks rebuild ghost positional
args, sort!(InsertionSort) compiles to wasm with zero offloads, and a
Base-functions differential corpus (gcd, binomial, isqrt, hash, mod,
fld, sum, generators, searchsortedfirst, hypot, floor, exp, sin, log,
string(n,base=16), sort!).

+++ WasmCodegen/src/reprs.jl
@@ -3,6 +3,13 @@
 # Scalars are stored in i32/i64/f32/f64. Sub-word integers (Int8/16, UInt8/16)
 # live in i32 with a normalization discipline: signed values are kept
 # sign-extended, unsigned values zero-extended; arithmetic renormalizes.
+#
+# Char is stored as its RAW 32-bit pattern (UTF-8 bytes left-justified) — the
+# exact same representation native Julia uses — both inside wasm and on the
+# wire. This makes bitcast/zext/trunc on Char identity operations and keeps
+# Base's codepoint decoding (e.g. `UInt32(::Char)`) bit-faithful. Do NOT store
+# the codepoint anywhere: mixing the two conventions silently corrupts every
+# Char that crosses the boundary.
 
 struct ScalarRepr
     vt::NumType      # wasm storage type
@@ -56,6 +63,7 @@ end
 """Convert a host value arriving as i32/i64/f32/f64 back to Julia type `T`."""
 function from_wire(@nospecialize(T), v)
     T === Bool && return v != 0
+    # Char travels as its raw bit pattern (must invert `to_wire` exactly)
     T === Char && return reinterpret(Char, UInt32(reinterpret(UInt32, Int32(v))))
     r = scalar_repr(T)
     r.isfloat && return T(v)
@@ -66,7 +74,7 @@ end
 """Convert a Julia scalar to what the wasm boundary expects."""
 function to_wire(@nospecialize(T), v)
     T === Bool && return Int32(v::Bool)
-    T === Char && return reinterpret(Int32, UInt32(v::Char))
+    T === Char && return reinterpret(Int32, v::Char)   # raw bits, not codepoint
     r = scalar_repr(T)
     r === nothing && throw(CompileError("unsupported boundary type $T"))
     r.isfloat && return T(v)
